requirements.txt
fastapi
requests
uvicorn
xmltodict
pycryptodome